Refactor DownloadAndUnzip
