+++ annotation-server/src/clinical_annotation/clinical_annotation.service.ts
@@ -2,11 +2,10 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { parse } from 'csv-parse';
 import * as fs from 'fs';
-import * as https from 'https';
 import * as os from 'os';
 import * as path from 'path';
 import { Repository } from 'typeorm';
-import * as yauzl from 'yauzl';
+import { downloadAndUnzip } from '../common/utils/download-unzip';
 import { ClinicalAnnotation } from './clinical_annotation.entity';
 
 @Injectable()
@@ -18,15 +17,13 @@ export class ClinicalAnnotationService {
 
   async syncAnnotations() {
     const url = 'https://s3.pgkb.org/data/clinicalAnnotations.zip';
-    const zipPath = path.join(os.tmpdir(), 'clinical_annotations.zip');
     const unzipTargetPath = path.join(os.tmpdir(), 'clinical_annotations');
     const annotationsTsvPath = path.join(
       unzipTargetPath,
       'clinical_annotations.tsv',
     );
 
-    await download(url, zipPath);
-    await unzip(zipPath, unzipTargetPath);
+    await downloadAndUnzip(url, unzipTargetPath);
     const annotations = await parseAnnotations(annotationsTsvPath);
 
     // Persist annotations in array
@@ -35,94 +32,6 @@ export class ClinicalAnnotationService {
     });
   }
 }
-
-const download = (url: string, filePath: string) => {
-  return new Promise<void>((resolve) => {
-    https.get(url, (res) => {
-      const writeStream = fs.createWriteStream(filePath);
-
-      res.pipe(writeStream);
-
-      writeStream.on('finish', () => {
-        writeStream.close();
-        resolve();
-      });
-    });
-  });
-};
-
-const unzip = (zipPath: string, unzipTargetPath: string) => {
-  return new Promise<void>((resolve, reject) => {
-    try {
-      // Create folder if not exists
-      if (!fs.existsSync(unzipTargetPath)) {
-        fs.mkdirSync(unzipTargetPath);
-      }
-
-      // Unzip file
-      yauzl.open(zipPath, { lazyEntries: true }, (error, zipfile) => {
-        if (error) {
-          zipfile.close();
-          reject(error);
-          return;
-        }
-
-        // Read first entry
-        zipfile.readEntry();
-
-        // Trigger next cycle, every time we read an entry
-        zipfile.on('entry', (entry) => {
-          // Directories
-          if (/\/$/.test(entry.fileName)) {
-            // If it is a directory, it needs to be created
-            const dirPath = path.join(unzipTargetPath, entry.fileName);
-            if (!fs.existsSync(dirPath)) {
-              fs.mkdirSync(dirPath);
-            }
-            zipfile.readEntry();
-          }
-          // Files
-          else {
-            zipfile.openReadStream(entry, (error, readStream) => {
-              if (error) {
-                zipfile.close();
-                reject(error);
-                return;
-              }
-
-              const filePath = path.join(unzipTargetPath, entry.fileName);
-              const file = fs.createWriteStream(filePath);
-              readStream.pipe(file);
-
-              file.on('finish', () => {
-                // Wait until the file is finished writing, then read the next entry.
-                file.close(() => {
-                  zipfile.readEntry();
-                });
-              });
-
-              file.on('error', (error) => {
-                zipfile.close();
-                reject(error);
-              });
-            });
-          }
-        });
-
-        zipfile.on('end', () => {
-          resolve();
-        });
-
-        zipfile.on('error', (error) => {
-          zipfile.close();
-          reject(error);
-        });
-      });
-    } catch (error) {
-      reject(error);
-    }
-  });
-};
 
 const parseAnnotations = (filePath: string) => {
   return new Promise<ClinicalAnnotation[]>((resolve, reject) => {
